Move to traits implementing the actual insertion logic
