+++ compiler/planner/mod.rs
@@ -19,4 +19,4 @@ pub mod program_plan;
 
 pub mod insert_planner;
 mod vertex;
-mod write_instructions;
+pub mod write_instructions;
