+++ executor/write/mod.rs
@@ -4,11 +4,5 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-mod annotation;
-mod constraints;
-mod expression;
-pub mod function;
-pub mod insert;
-pub mod literal;
-pub mod match_;
-pub mod tokens;
+pub mod insert_executor;
+pub(crate) mod write_instruction;
